requirements.txt
tensorflow-macos>=2.16.0
numpy>=1.24.0
Flask>=3.0.0
Pillow>=10.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
tqdm>=4.65.0
beautifulsoup4>=4.12.0
requests>=2.31.0 